+++ next.config.ts
@@ -3,9 +3,7 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   serverExternalPackages: [
     "playwright",
-    "playwright-extra",
     "playwright-core",
-    "puppeteer-extra-plugin-stealth",
     "@axe-core/playwright",
   ],
 };